+++ src/main/java/vendingmachine/Controller/InitController.java
@@ -1,12 +1,10 @@
 package vendingmachine.Controller;
 
 import vendingmachine.Service.InitHoldingCoinsService;
-import vendingmachine.Service.ValidatorService;
 import vendingmachine.View.InputView;
 import vendingmachine.View.OutputView;
 
 public class InitController {
-    ValidatorService validatorService = new ValidatorService();
     InitHoldingCoinsService holdingCoinsService = new InitHoldingCoinsService();
 
     public void initVendingMachine() {
@@ -16,7 +14,7 @@ public class InitController {
     private void initHoldingCoins() {
         try {
             String holdingAmount = InputView.getHoldingAmount();
-            validatorService.isValidHoldingAmount(holdingAmount);
+            holdingCoinsService.isValidHoldingAmount(holdingAmount);
             holdingCoinsService.setHoldingCoins(Integer.parseInt(holdingAmount));
             OutputView.printHoldingCoins();
         } catch (IllegalArgumentException e) {
